XECS-1431 - Patch from Carolyn Beeton
Fix unit test code for 64-bit, cast unsigned constant as size_t.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13235 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXsupervisor/src/test/SipxResourceUseTest.cpp
@@ -234,7 +234,7 @@ public:
          fileResource->appendDescription(description);
          ASSERT_STR_EQUAL("file '/etc/sipxpbx/common.xml'",description.data());
 
-         CPPUNIT_ASSERT_EQUAL(2U, fileResource->mUsedBy.entries());
+         CPPUNIT_ASSERT_EQUAL((size_t)2, fileResource->mUsedBy.entries());
       };
    
 };
